Change: switiching from rocket to hyper
Instead of asking much is the fish we went hyper, hyper.

Replaces the rocket implementation to hyper so that we have more fine
granular control.

Since the scanner-api definition is rather simple there are not that
many endpoints to consider which makes it feasible to user hyper
directly.

There are still a lot of things open to do:
- fix crucial bug that it does neither exit nor recover on panic
- write more tests
- document the code